Add function to generate matching block dictionary
As before this new function generate_matching_blocks_dict() is not
called, but is unit tested to ensure that it works correctly.

Given a list of MD5 hashes this function generate a dictionary where
each key is the MD5 hash and the value is the list of data blocks with
that hash.