Add function to remove non-duplicate items from matching hashes
A unique MD5 hash of a block says the block is unique and therefore is
not within the overlapping range.  So remove it from the matching hashes
dictionary.

Also remove MD5 hashes with three or more listed matching blocks as we
are looking for blocks with two copies in the overlapping range.  As the
FIXME code comment says it is unknown whether using offsets between
groups of three or more blocks with the same MD5 hash will add to the
signal or noise.  Investigate.

Coding references:
* Iterating over dictionaries using 'for' loops
  https://stackoverflow.com/questions/3294889/iterating-over-dictionaries-using-for-loops
* Delete an element from a dictionary
  https://stackoverflow.com/questions/5844672/delete-an-element-from-a-dictionary