Add function to compute the offset blocks dictionary
The offset blocks dictionary maps from the offset between pairs of
blocks sharing the same MD5 hash to the list of the first of such pairs
with that particular offset.  Namely this dictionary:

    offset_blocks = {7: [1], 3: [10, 11, 12]}

Read this dictionary as saying block [1] has a partner with the same MD5
hash at offset 7, namely block [8], and blocks [10, 11, 12] have
partners with the same MD5 hash at offset 3, namely block [13, 14, 15]
respectively.

This is the next step in being able to identify the size of the
overlapping range and then the blocks involved.