Also keep groups of 3 and 4 blocks with the same hash
A specially crafted test file system image (test-hashes-t1.txt.xz) with
2 copies of a set of files and the overlapping range in the middle
contained lots of groups of 3 blocks with the same MD5 hash.  This
strongly added to the signal in finding the overlapping range.  It is
considered unlikely that a user's file system will have more than 3
copies of files exactly in the overlapping range.  Therefore drop groups
of more than 4 blocks with the same MD5 hash.