Add function to compute candidate overlap ranges
Takes the offset blocks dictionary and computes the list of possible
candidate overlapping data ranges.  For each candidate overlapping size
it finds the range of blocks and generates a rank as size of range
divided by size of offset.  A rank of 1.0 indicates the matching range
exactly equals the offset and a valid overlapping data range has been
found.