Add integration test f18089
GParted Forum user reported [1] that GParted failed when performing a
shrink and move of a file system when it encountered a hard drive error.
This left the file system partially copied with an overlapping range in
the middle.  Finding the overlapping range was the impetus which
ultimately lead to writing this program.  Add dumped hashes generated
from this case and use as an integration test.

There are now multiple make targets run as:
    $ make                      Runs default "check" target
    $ make check                Runs "unit-tests" target
    $ make unit-tests
    $ make integration-tests
    $ make all-tests

[1] [SOLVED] Unallocated space and corrupted NTFS after resize
    http://gparted-forum.surf4.info/viewtopic.php?id=18089

[2] How to invoke pytest, Run tests by keyword expressions
    https://docs.pytest.org/en/7.1.x/how-to/usage.html